Split integration tests common module into submodules.
